Revert "Upgrade of dashboard page"
This reverts commit 33dd3cfda1c64f3b2739979e922e0840535e0b8c.
