+++ extracted/en.json
@@ -116,14 +116,9 @@
   "coaches.heading": "Coaches",
   "controls.back": "Back",
   "controls.next": "Next",
-  "dashboard.actions.empty.action": "Start private session",
-  "dashboard.actions.empty.perex": "Go through a session to set action points for your goal",
-  "dashboard.actions.empty.title": "No actions",
   "dashboard.cards.ai.empty.perex": "Unlock deeper insights by revealing your Strengths and Values. With added feedback, the insights become even sharper.",
   "dashboard.cards.ai.empty.steps.explore.title": "Explore uniqueness & get growth tips",
-  "dashboard.cards.ai.empty.steps.strengths.button": "Find my strengths",
   "dashboard.cards.ai.empty.steps.strengths.title": "Assess strengths",
-  "dashboard.cards.ai.empty.steps.values.button": "Set my values",
   "dashboard.cards.ai.empty.steps.values.title": "Define values",
   "dashboard.cards.ai.empty.title": "Insights and daily tips",
   "dashboard.cards.ai.insights.animal-spirit.title": "Animal Spirit Guide - Discover the animal that resonates with your leadership persona",
@@ -138,16 +133,11 @@
   "dashboard.cards.notes.placeholder.empty": "Type something",
   "dashboard.cards.notes.title": "My notes",
   "dashboard.cards.sessions.title.filled": "My areas of development",
-  "dashboard.cards.strengths.perex": "Understand your top talents to focus on what you excel at.",
   "dashboard.cards.strengths.title.empty": "Find my strengths",
   "dashboard.cards.strengths.title.filled": "My strengths",
-  "dashboard.cards.values.perex": "Your values are a central part of who you are.",
   "dashboard.cards.values.title.empty": "Set my values",
   "dashboard.cards.values.title.filled": "My values",
   "dashboard.header": "Hello, {user}!",
-  "dashboard.learn.heading": "Learn",
-  "dashboard.my-leader-profile": "My leader profile",
-  "dashboard.perex": "Sapiente laboriosam doloribus cumque tempora quibusdam et explicabo facilis.",
   "dashboard.rightmenu.actions.heading": "Actions",
   "dashboard.rightmenu.actions.perex.empty": "Once you set a goal, actions will apear here",
   "dashboard.rightmenu.actions.set-area": "Set area for my development",
@@ -161,12 +151,6 @@
   "dashboard.section-1.perex": "Before you set where you are heading it's good to know from where you start",
   "dashboard.section-2.heading": "Who to become",
   "dashboard.section-2.perex": "Become a better leader and here is how you get there.",
-  "dashboard.unlock-steps.perex": "Once you pass the assessments, your strengths and values will appear here",
-  "dashboard.upcoming.empty.action.schedule-session": "Schedule session",
-  "dashboard.upcoming.empty.action.select-coach": "Select coach",
-  "dashboard.upcoming.empty.perex": "Sessions with a coach will appear here",
-  "dashboard.upcoming.empty.title": "No upcoming sessions",
-  "dashboard.upcoming.heading": "Upcoming sessions",
   "dict.areas.1": "Become an active listener",
   "dict.areas.2": "Become more efficient",
   "dict.areas.3": "Show appreciation, recognition and empathy for your team",
@@ -757,7 +741,6 @@
   "general.new-message-from": "New message from {from}",
   "general.notifications": "Notifications",
   "general.select.placeholder": "Select from the list",
-  "general.today": "Today",
   "google.back-to-dashboard": "Back to dashboard",
   "google.error.description": "Unfortunately, we were unable to integrate your Google Calendar with the TopLeader platform at this time. Please try again later. If the issue persists, do not hesitate to contact us at {emailLink} for assistance.",
   "google.error.title": "Failed",
