+++ package.json
@@ -7,7 +7,7 @@
     "@emotion/styled": "latest",
     "@mui/icons-material": "^5.10.3",
     "@mui/lab": "^5.0.0-alpha.102",
-    "@mui/material": "5.16.6",
+    "@mui/material": "5",
     "@mui/x-date-pickers": "^6.16.2",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
@@ -39,7 +39,9 @@
     "build-github:qa": "DISABLE_ESLINT_PLUGIN=true REACT_APP_GIT_SHA=`git rev-parse --short HEAD` REACT_APP_ENV=QA react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "extract-translations": "formatjs extract 'src/**/*.js*' --ignore='src/translations' --out-file extracted/en.json --id-interpolation-pattern '[sha512:contenthash:base64:6]' --format scripts/extract-translations-formatter.js",
+    "extract-translations-long": "formatjs extract 'src/**/*.js*' --ignore='src/translations' --out-file extracted/en-with-descriptions.json --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
+    "extract-translations-short": "formatjs extract 'src/**/*.js*' --ignore='src/translations' --out-file extracted/en.json --id-interpolation-pattern '[sha512:contenthash:base64:6]' --format scripts/extract-translations-formatter.js",
+    "extract-translations": "npm run extract-translations-short ; npm run extract-translations-long",
     "compile": "formatjs compile",
     "translations-match-csv": "node ./scripts/translations-csv-converter.js",
     "generate-translations": "npm run extract-translations && npm run translations-match-csv",
