+++ src/components/Icon.js
@@ -36,7 +36,6 @@ import {
   KeyboardDoubleArrowRight,
   Lightbulb,
   Lock,
-  School,
   Login,
   LogoutOutlined,
   Mail,
@@ -58,9 +57,6 @@ import {
   LinkedIn,
   StopCircle,
   Sync,
-  CalendarMonth,
-  PlayCircleOutline,
-  MenuBookOutlined,
 } from "@mui/icons-material";
 import * as React from "react";
 
@@ -101,7 +97,6 @@ const Icons = {
   KeyboardDoubleArrowRight,
   Lightbulb,
   Lock,
-  School,
   Login,
   LogoutOutlined,
   Mail,
@@ -123,9 +118,6 @@ const Icons = {
   LinkedIn,
   StopCircle,
   Sync,
-  CalendarMonth,
-  PlayCircleOutline,
-  MenuBookOutlined,
 };
 
 export const ICON_NAMES = Object.keys(Icons);
