+++ src/components/ScrollableRightMenu.js
@@ -7,7 +7,6 @@ export const ScrollableRightMenu = ({
   children,
   buttonProps,
   sx = {},
-  wrapperSx = {},
 }) => {
   return (
     <ErrorBoundary>
@@ -33,7 +32,6 @@ export const ScrollableRightMenu = ({
             display: "flex",
             flexDirection: "column",
             flexWrap: "nowrap",
-            ...wrapperSx,
           }}
         >
           {children}
