+++ src/features/Messages/Messages.page.js
@@ -90,7 +90,7 @@ const stringToColor = (string) => {
   /* eslint-enable no-bitwise */
 };
 
-export const UserAvatar = ({
+const UserAvatar = ({
   username: usernameProp,
   fullName: fullNameProp,
   sx = {},
