+++ src/translations/cs.json
@@ -672,15 +672,12 @@
   "main-menu.items.settings": "Nastavení",
   "settings.heading": "Nastavení",
   "main-menu.items.help": "Pomoc",
-  "dashboard.cards.ai.empty.steps.values.button": "Určit si své hodnoty",
   "dashboard.cards.values.title.empty": "Určit si své hodnoty",
   "dashboard.cards.notes.title": "Moje poznámky",
   "dashboard.section-2.heading": "Kým se stát",
   "dashboard.section-2.perex": "Staňte se lepším leaderem a tímto se tam dostanete.",
   "dashboard.rightmenu.title": "Moje leadership cesta",
   "dashboard.rightmenu.upcoming.title.empty": "Není nastaven osobní rozvoj",
-  "dashboard.upcoming.empty.title": "Není nastaven osobní rozvoj",
-  "dashboard.upcoming.empty.perex": "Osobní rozvoj s koučem se zobrazí zde",
   "dashboard.rightmenu.actions.heading": "Aktivity",
   "dashboard.rightmenu.actions.title.empty": "Aktivity nejsou určeny",
   "dashboard.rightmenu.actions.perex.empty": "Jakmile si stanovíte cíl, aktivity se zobrazí zde",
@@ -1014,7 +1011,6 @@
   "dict.feedback.question.efficiency.improve-processes.label": "Zlepšuje osoba procesy, aby byly efektivnější?",
   "dict.validation.required": "Požadované",
   "feedback.create.stats.required": "Požadované",
-  "dashboard.cards.ai.empty.steps.strengths.button": "Poznat své silné stránky",
   "settings.tabs.credits.label": "Kredity",
   "clients.add": "Doplnit člena",
   "settings.admin.add": "Doplnit člena",
