+++ src/translations/de.json
@@ -672,15 +672,12 @@
   "main-menu.items.settings": "Einstellungen",
   "settings.heading": "Einstellungen",
   "main-menu.items.help": "Hilfe",
-  "dashboard.cards.ai.empty.steps.values.button": "Meine Werte bestimmen",
   "dashboard.cards.values.title.empty": "Meine Werte bestimmen",
   "dashboard.cards.notes.title": "Meine Notizen",
   "dashboard.section-2.heading": "Wo möchte ich hin",
   "dashboard.section-2.perex": "Werden Sie eine bessere Führungskraft und hier finden Sie den Weg dazu",
   "dashboard.rightmenu.title": "Meine Führungsreise",
   "dashboard.rightmenu.upcoming.title.empty": "Keine Persönlichkeitsentwicklung vorhanden",
-  "dashboard.upcoming.empty.title": "Keine Persönlichkeitsentwicklung vorhanden",
-  "dashboard.upcoming.empty.perex": "Persönlichkeitsentwicklung mit einem Couch werden hier erscheinen",
   "dashboard.rightmenu.actions.heading": "Aktionen",
   "dashboard.rightmenu.actions.title.empty": "Keine Aktionen festgelegt",
   "dashboard.rightmenu.actions.perex.empty": "Sobald Sie ein Ziel festgelegt haben, werden Aktionen hier angezeigt",
@@ -1014,7 +1011,6 @@
   "dict.feedback.question.efficiency.improve-processes.label": "Verbessert der Mitarbeiter Prozesse, um sie effektiver zu gestalten?",
   "dict.validation.required": "Erforderlich",
   "feedback.create.stats.required": "Erforderlich",
-  "dashboard.cards.ai.empty.steps.strengths.button": "Meine Stärken finden",
   "settings.tabs.credits.label": "Guthaben",
   "clients.add": "Mitglied hinzufügen",
   "settings.admin.add": "Mitglied hinzufügen",
