+++ src/translations/en.json
@@ -672,15 +672,12 @@
   "main-menu.items.settings": "Settings",
   "settings.heading": "Settings",
   "main-menu.items.help": "Help",
-  "dashboard.cards.ai.empty.steps.values.button": "Set my values",
   "dashboard.cards.values.title.empty": "Set my values",
   "dashboard.cards.notes.title": "My notes",
   "dashboard.section-2.heading": "Who to become",
   "dashboard.section-2.perex": "Become a better leader and here is how you get there.",
   "dashboard.rightmenu.title": "My leadership journey",
   "dashboard.rightmenu.upcoming.title.empty": "No upcoming sessions",
-  "dashboard.upcoming.empty.title": "No upcoming sessions",
-  "dashboard.upcoming.empty.perex": "Sessions with a coach will appear here",
   "dashboard.rightmenu.actions.heading": "Actions",
   "dashboard.rightmenu.actions.title.empty": "No actions set",
   "dashboard.rightmenu.actions.perex.empty": "Once you set a goal, actions will apear here",
@@ -1014,7 +1011,6 @@
   "dict.feedback.question.efficiency.improve-processes.label": "Does the employee improve processes to make them more effective?",
   "dict.validation.required": "Required",
   "feedback.create.stats.required": "Required",
-  "dashboard.cards.ai.empty.steps.strengths.button": "Find my strengths",
   "settings.tabs.credits.label": "Credits",
   "clients.add": "Add member",
   "settings.admin.add": "Add member",
