+++ src/translations/fr.json
@@ -672,15 +672,12 @@
   "main-menu.items.settings": "Paramètres",
   "settings.heading": "Paramètres",
   "main-menu.items.help": "Aide",
-  "dashboard.cards.ai.empty.steps.values.button": "Déterminer mes valeurs",
   "dashboard.cards.values.title.empty": "Déterminer mes valeurs",
   "dashboard.cards.notes.title": "Mes notes",
   "dashboard.section-2.heading": "Qui devenir?",
   "dashboard.section-2.perex": "Devenez un meilleur leader et voici comment vous y parviendrez.",
   "dashboard.rightmenu.title": "Mon chemin vers le leadership",
   "dashboard.rightmenu.upcoming.title.empty": "Aucunes sessions à venir.",
-  "dashboard.upcoming.empty.title": "Aucunes sessions à venir.",
-  "dashboard.upcoming.empty.perex": "Les sessions avec un coach/formateur vont être visibles ici.",
   "dashboard.rightmenu.actions.heading": "Les actions",
   "dashboard.rightmenu.actions.title.empty": "Aucune action définie.",
   "dashboard.rightmenu.actions.perex.empty": "Dès lors que vous établissez un objectif, les actions/tâches apparaîtront ici.",
@@ -1014,7 +1011,6 @@
   "dict.feedback.question.efficiency.improve-processes.label": "L'employé améliore-t-il les processus pour les rendre plus efficaces?",
   "dict.validation.required": "Requis",
   "feedback.create.stats.required": "Requis",
-  "dashboard.cards.ai.empty.steps.strengths.button": "Connaissez mes points forts",
   "settings.tabs.credits.label": "Crédits",
   "clients.add": "Ajouter un membre",
   "settings.admin.add": "Ajouter un membre",
